Moves mocks-reading function to a separate class
