+++ WooCommerce/src/androidTest/kotlin/com/woocommerce/android/screenshots/util/MocksReader.kt
@@ -1,0 +1,23 @@
+package com.woocommerce.android.screenshots.util
+
+import androidx.test.platform.app.InstrumentationRegistry
+import org.json.JSONArray
+import org.json.JSONObject
+
+class MocksReader {
+
+    fun readAllReviewsToArray(): JSONArray {
+        val reviewsWireMockFileName = "mocks/mappings/jetpack-blogs/wc/reviews/products_reviews_all.json"
+        val reviewsWireMockString = this.readAssetsFile(reviewsWireMockFileName)
+        val reviewsWireMockJSON = JSONObject(reviewsWireMockString)
+        return reviewsWireMockJSON
+            .getJSONObject("response")
+            .getJSONObject("jsonBody")
+            .getJSONArray("data")
+    }
+
+    private fun readAssetsFile(fileName: String): String {
+        val appContext = InstrumentationRegistry.getInstrumentation().context
+        return appContext.assets.open(fileName).bufferedReader().use { it.readText() }
+    }
+}
